fix: dashboard has been moved to pages
